\name{dimensionreduce}
\alias{dimensionreduce}
\title{Reduce dimensions and return the best by testing accuracy}
\usage{
dimensionreduce(data, label)
}
\arguments{
data

Only contains the data columns that you want to reduce their dimensions, no label! And make sure it's in dataframe type.

label

Only contains the label column, which means it should be only one column, no data! And make sure its rows are matched with data, and make sure it's in dataframe type.
}
\description{
For this function, you need to input 
}
